requirements.txt
itsdangerous==1.1.0
Click==7.0
numpy==1.16.2
Flask==1.0.2
Flask-RESTful==0.3.7
flask-restplus==0.13.0
h5py==2.9.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
pandas==0.24.1
pickleshare==0.7.5
requests==2.20.0
requests-oauthlib==1.2.0
tensorboard==1.12.2
tensorflow==1.12.0
virtualenv==16.1.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
flask-restplus==0.13.0
flask-restplus==0.13.0
Pillow==6.1.0
